The get_valid_word(words): function
Description: if I print out words from the words.py file using print(words) I will get all the words ni that file. The first step to get the computer to play hangman with the users is that the computer has to figure out a word for the players to guess. The player has to constanly guess and pick a word that is valid that can be used in hangman that is why get_valid_word function is used. random.choice makes the program/computer pick any random word from words.py file and store it in the variable word after the function passes the words to it from its parameter.