The while loop
Description: The while loop continues to loop between all the words until there is no more space or -. That is how the compupter will choose a word out of the 1000s of words in the file. As long as there is - or ' ' in 'word' keep choosing this word once it stops iterating we will get a word that no longer has a - or space in it. Once it finds the word it returns it