The hangman() function
Description: We need to keep track of what is a valid letter, what letters we guessed and which letters in the word we correctly guessed. string is imported in order to use string.ascii_uppercase which turns letters into uppercase. set() method is used to convert any of the iterable to sequence of iterable elements with dintinct elements, commonly called Set. used_letters is to keep track of what the user has guessed.