Getting user input
Description: What we can do is ask for user input User the user_input as input. if user_letter in the alphabet is a valid character that i havent used yet i am going to add to user set. else if user_letter in usedPP_letters jsut let them know that you already used this character. else if they type something that is not in the alphabet and is not in the used letter that they guessed that means they typed in the wrong character so we just let them know they typed in an invalid character