While loop
Now that we can get the user input we want the user to be able to keep guessing until they get the word in order to make that happen we are going to use a while loop so the user can keep on guessing until he guess the right word. we are removing a word letter from set words which we haven't seen yet. The condition that needs to be satisfied after the user gets all the word is the length or 'len' of word letter is = 0. so this while loop will continue to loop and substract used letters in alphabet.