Intro
Description: Some of the words from words. py have ' ' or spaces and - or dashes in between the words, so a function needs to be created that has a while loop that continues to iterate while there is - or space in between the words. This function takes care of the words with spaces and dashes in between them, so it will continue to pick words or iterate from the list until there isn't any with - or space in between those words.