hangman() function keep track of what is a valid letter
Description: We need to keep track of the words we guessed and which words we correctly guessed. get_valid_words is stored in word, word_letters = set(word).
 A Python set is an unordered collection of data with no duplicate elements. A set supports operations like union, intersection, or difference, known from Mathematics.
alphabet variable gets a set of ascii_uppercase strings in order to keep all the letters uppercase and reduce errors.
So word uses the get_valid_word function to get valid word from list of words then the word varible is used to remove duplicate words and set those words which is then stored in the word_letters varible. then string is imported and the string.ascii_uppercase is used and stored in alphabet variable. Finally variable used_letters is used to keep track of the words that the user guessed.