Getter user input
Description: First thing we do is ask for user input directly by instructing them to Guess a Letter, once they enter a letter it will be converted to uppercase using the upper() method then it is stored in the 'user_letter' variable.
A lower case a in python is different from an uppercase A, if you try to test equality between those two strings it won't be equal so .upper() method is used.
If user letter in alphabet has not been used yet add user's letter or user_letter to used_letters. The add() method is used to add a letter to a variable.
If user letter is in the variable word_letters remove user's letter or user_letter from word_letters variable which has a saved set in it.
else if the user's letter is in in a set of letters called used_letters print and let the user know that he as already used that character he needs to give it another try.
else in any other case print 'Invalid character. Try again '